# osdu/specs/production-dspdm/swagger-api.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Production DDMS(core) Service",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/{MID_PATH}",
    "description" : "{HOST}/{MID_PATH}/{API_PATH}",
    "variables" : {
      "MID_PATH" : {
        "default" : ""
      }
    }
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  }, {
    "partationId" : [ ]
  } ],
  "paths" : {
...
    "/v1/spatialdata/spatialdescriptions/lines/points/{id}" : {
      "put" : {
        "tags" : [ "Spatial Line Data" ],
        "summary" : "Update spatial point data with given point identifier.",
        "description" : "This API is used to update spatial line point data with given point identifier.",
        "operationId" : "updateSpatialLinePoint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Spatial line point identifier.",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GeneratedSpatialPoint"
              },
              "example" : {
                "depth" : 123,
                "xCoordinate" : 123.18,
                "yCoordinate" : 123.12,
                "remark" : "Updated point"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "The data is updated.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UpdatedSpatialLinePoint"
                },
                "example" : {
                  "code" : 200,
                  "version" : "v1",
                  "message" : "Update successfully",
                  "data" : {
                    "spLinePointID" : 165239,
                    "pointSeqNo" : 5,
                    "xCoordinate" : 123.18,
                    "yCoordinate" : 123.12,
                    "depth" : 123,
                    "remark" : "Updated point"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Missing mandatory data."
          },
          "401" : {
            "description" : "Not authorized!"
          },
          "403" : {
            "description" : "Forbidden!"
          },
          "404" : {
            "description" : "No data is found for given identifier."
          }
        }
      },
      "delete" : {
        "tags" : [ "Spatial Line Data" ],
        "summary" : "Delete spatial line data with given line point identifier.",
        "description" : "This API is used to delete a line data with given line point identifier.",
        "operationId" : "deleteSpatialLinePoint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Spatial line point identifier.",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The data is deleted.",
            "content" : {
              "application/json" : {
                "example" : {
                  "code" : 200,
                  "version" : "v1",
                  "message" : "Delete successfully",
                  "data" : null
                }
              }
            }
          },
          "400" : {
            "description" : "Missing mandatory data."
          },
          "401" : {
            "description" : "Not authorized!"
          },
          "403" : {
            "description" : "Forbidden!"
          },
          "404" : {
            "description" : "No data is found for given identifier."
          }
        }
      }
    },
    "/v1/spatialdata/spatialdescriptions/points/{id}" : {
      "put" : {
        "tags" : [ "Spatial Point Data" ],
        "summary" : "Update spatial point data with given point identifier.",
        "description" : "This API is used to update spatial point data with given point identifier.",
        "operationId" : "updateSpatialPoint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Spatial point identifier.",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GeneratedSpatialPoint"
              },
              "example" : {
                "xCoordinate" : 111.18,
                "yCoordinate" : 222.12,
                "depth" : 333,
                "remark" : "Updated spatial point"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "The data is updated.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FullSpatialPoint"
                },
                "example" : {
                  "code" : 200,
                  "version" : "v1",
                  "message" : "Update successfully",
                  "data" : {
                    "spPointID" : 65,
                    "spatialDescriptionID" : 163,
                    "xCoordinate" : 111.18,
                    "yCoordinate" : 222.12,
                    "depth" : 333,
                    "remark" : "Updated spatial point"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Missing mandatory data."
          },
          "401" : {
            "description" : "Not authorized!"
          },
          "403" : {
            "description" : "Forbidden!"
          },
          "404" : {
            "description" : "No data is found for given identifier."
          }
        }
      },
      "delete" : {
        "tags" : [ "Spatial Point Data" ],
        "summary" : "Delete spatial point data with given point identifier.",
        "description" : "This API is used to delete a specific point data with given point identifier.",
        "operationId" : "deleteSpatialPoint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Spatial point identifier.",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The data is deleted.",
            "content" : {
              "application/json" : {
                "example" : {
                  "code" : 200,
                  "version" : "v1",
                  "message" : "Delete successfully",
                  "data" : null
                }
              }
            }
          },
          "400" : {
            "description" : "Missing mandatory data."
          },
          "401" : {
            "description" : "Not authorized!"
          },
          "403" : {
            "description" : "Forbidden!"
          },
          "404" : {
            "description" : "No data is found for given identifier."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FullSpatialPoint" : {
        "type" : "object",
        "properties" : {
          "spPointID" : {
            "type" : "integer",
            "description" : "The identifier of the spatial point.",
            "format" : "int32",
            "example" : 3001
          },
          "xCoordinate" : {
            "type" : "number",
            "description" : "X location, which can hold a geographical longitude or projected X depending on the value of CRS.",
            "format" : "double",
            "example" : 11579
          },
          "yCoordinate" : {
            "type" : "number",
            "description" : "Y location, which can hold a geographical longitude or projected Y depending on the value of CRS.",
            "format" : "double",
            "example" : 12056
          },
          "depth" : {
            "type" : "number",
            "description" : "The Depth of a position.",
            "format" : "double",
            "example" : 300
          },
          "remark" : {
            "type" : "string",
            "description" : "Sub-surface location of Equipment 1.",
            "example" : "Narrative description of this row of data."
          }
        },
        "description" : "Spatial point response schema."
      },
...
      "GeneratedSpatialPoint" : {
        "type" : "object",
        "properties" : {
          "xCoordinate" : {
            "type" : "number",
            "description" : "X location, which can hold a geographical longitude or projected X depending on the value of CRS.",
            "format" : "double",
            "example" : 11579
          },
          "yCoordinate" : {
            "type" : "number",
            "description" : "Y location, which can hold a geographical longitude or projected Y depending on the value of CRS.",
            "format" : "double",
            "example" : 12056
          },
          "depth" : {
            "type" : "number",
            "description" : "The Depth of a position.",
            "format" : "double",
            "example" : 300
          },
          "remark" : {
            "type" : "string",
            "description" : "Sub-surface location of Equipment 1.",
            "example" : "Narrative description of this row of data."
          }
        },
        "description" : "Spatial point data. Required when sp desc type is point."
      },
...
      "UpdatedSpatialLinePoint" : {
        "type" : "object",
        "properties" : {
          "spLinePointID" : {
            "type" : "integer",
            "description" : "The identifier of the point.",
            "format" : "int32",
            "example" : 3001
          },
          "pointSeqNo" : {
            "type" : "integer",
            "description" : "Point sequence number of each point.",
            "format" : "int32",
            "example" : 1
          },
          "xCoordinate" : {
            "type" : "number",
            "description" : "X location, which can hold a geographical longitude or projected X depending on the value of CRS.",
            "format" : "double",
            "example" : 11579
          },
          "yCoordinate" : {
            "type" : "number",
            "description" : "Y location, which can hold a geographical longitude or projected Y depending on the value of CRS.",
            "format" : "double",
            "example" : 12056
          },
          "depth" : {
            "type" : "number",
            "description" : "The Depth of a position.",
            "format" : "double",
            "example" : 300
          },
          "remark" : {
            "type" : "string",
            "description" : "Narrative description.",
            "example" : "Narrative description of this row of data."
          }
        },
        "description" : "Point data of spatial line."
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "partationId" : {
        "type" : "apiKey",
        "name" : "data-partition-id",
        "in" : "header"
      }
    }
  }
}